+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## current
 - Edit lifetime stats on new player/goalie creation
+- Sort goalies by minutes played
 
 ## 0.11.0
 - Added active flag to players/goalies