+++ ChangeLog.md
@@ -4,6 +4,7 @@
 - Edit lifetime stats on new player/goalie creation
 - Sort goalies by minutes played
 - Subsort players by lifetime points
+- Changed wording on edit menus
 
 ## 0.11.0
 - Added active flag to players/goalies